update lpb structure
git-svn-id: https://ladd00.triumf.ca/svn/fgddaq/trunk/slowcontrol/c8051@966 6c71da66-6f2f-0410-bb3c-c2389e3dc2a8

+++ loader.h
@@ -233,7 +233,7 @@ struct EEPAGE xdata eepage={
 //
 // Global to ALL boards
 // Macro EEPAGE size independently of the BOARD
-#define PAGE_SIZE  sizeof(eepage)
+#define PAGE_SIZE (unsigned int) sizeof(eepage)
 // EEPROM page for loader confirmation
 struct EEPAGE xdata eepage0;
 struct EEPAGE xdata eepageWp;
